feat: 홈 화면 썸네일 openGraph image 상대주소 대응
